# BackendServer/demo/api.yaml
openapi: 3.0.0
info:
    title: FDS
    description: |
        This API allows to communicate with FDS (Food Delivery System) server.
    version: 0.0.8
paths:
...
    /product/{id}:
        get:
            summary: Get Product Details
            description: Retrieve details for a specific item by providing its ID
            operationId: getProduct
            parameters:
                - name: id
                  in: path
                  description: ID of the item to retrieve details for
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/foodItemFull'
                "404":
                    description: File Not Found
                "401":
                    description: Unauthorized
    /product/{id}/available:
        get:
            summary: Check Availability
            description: Check the availability of a specific item by providing its ID and quantity count
            operationId: checkProductAvailable
            parameters:
                - name: id
                  in: path
                  description: ID of the item to check availability for
                  required: true
                  schema:
                      type: string
                - name: count
                  in: query
                  description: Quantity count to check availability for
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: boolean
                                description: Indicates if the specified quantity is available
                "401":
                    description: Unauthorized
    /file/{fileId}:
        get:
            summary: Get file by file ID
            description: Retrieve a file, typically an image, based on the provided file ID.
            operationId: getFile
            parameters:
                - name: fileId
                  in: path  
                  description: ID of the file to retrieve
                  required: true
                  schema:
                      $ref: '#/components/schemas/fileId'
                - name: sessionId
                  in: header
                  required: false
                  schema:
                      $ref: '#/components/schemas/sessionId'
            responses:
                "200":
                    description: OK
                    content:
                        image/*:
                            schema:
                                type: multipartFile
                "401":
                    description: "Unauthorized"
                "404":
                    description: "File Not Found"
    /file:
        put:
            summary: Upload a file
            description: Upload an image to server for referencing elsewhere.
            operationId: putFile
            parameters:
                - name: sessionId
                  in: header
                  required: true
                  schema:
                      $ref: '#/components/schemas/sessionId'
            requestBody:
                required: true
                content:
                    image/*:
                        schema:
                            type: multipartFile # Assuming a base64-encoded image string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fileId'
                "401":
                    description: Unauthorized (if user not valid to upload file)
...
components:
    schemas:
        sessionId:
            type: string
            maxLength: 40
            minLength: 40
...
        fileId:
            type: string
            maxLength: 40
            minLength: 40
...
        foodItemFull:
            type: object
            properties:
                itemId:
                    type: string
                    description: ID of the item
                itemName:
                    type: string
                    description: Name of the item
                thumbnailPicture:
                    type: string
                    description: 64bit encoded thumbnail picture
                price:
                    type: number
                    description: Price of the item
                vendorName:
                    type: string
                    description: Name of the vendor
                vendorLocation:
                    type: string
                    description: Location of the vendor
                starRating:
                    type: number
                    description: Star rating of the item
                isAvailable:
                    type: boolean
                    description: Indicates if the item is available
                maxQuantity:
                    type: integer
                    description: Maximum quantity available
                imageUrls:
                    type: array
                    description: URLs of additional images
                    items:
                        type: string
                tags:
                    type: array
                    description: tags related to food item
                    items:
                        type: string
...
servers:
    - url: http://13.233.99.87:8080